refactor(cli): address review feedback
- Use inclusive shebang for multi-runtime support (bun/deno/node fallback)
- Replace custom parser with Node.js util.parseArgs (compatible with all runtimes)
- Simplify file loader by removing external configuration support
- Consolidate CLI files into single src/cli.ts for better maintainability
- Use direct 'srvx' import instead of runtime-specific adapter paths
- Remove separate cli/ directory and integrate all functionality

Addresses feedback from PR #91 review.
